requirement.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2022.12.7
cffi==1.17.1
charset-normalizer==3.1.0
cryptography==43.0.3
diff-match-patch==20241021
Django==5.0.1
django-crispy-forms==1.11.2
django-import-export==4.1.1
django-reset-migrations==0.4.0
django-utils-six==2.0
djangorestframework==3.15.2
idna==3.4
Pillow==9.5.0
pycparser==2.22
PyMySQL==1.0.2
pytz==2021.1
requests==2.28.2
six==1.16.0
soupsieve==2.6
sqlparse==0.4.1
tablib==3.5.0
tzdata==2024.2
urllib3==1.26.15
